+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 1.1.6 (...)
 - implement `PdoAuth` method to check if user exists
+- implement `InputValidation::voucherCode`
 
 ## 1.1.5 (2018-02-19)
 - implement method to update password in `PdoAuth`